requirements.txt
alembic==1.15.2
aniso8601==10.0.0
asttokens==2.4.1
attrs==25.3.0
autocommand==2.2.2
blinker==1.6.2
branca==0.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
cheroot==10.0.0
CherryPy==18.9.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
dnspython==2.7.0
dominate==2.9.1
email-validator==2.0.0
executing==2.0.1
filelock==3.17.0
Flask==2.2.5
Flask-Admin==1.6.1
Flask-Bootstrap==3.3.7.1
Flask-Dance==7.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Minify==0.37
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
folium==0.19.2
fonttools==4.54.1
fsspec==2024.12.0
geopandas==1.0.1
greenlet==3.1.1
gunicorn==20.1.0
htmlmin==0.1.12
htmlminf==0.1.13
idna==3.10
inflect==7.2.0
ipython==8.23.0
itsdangerous==2.2.0
jaraco.collections==5.0.0
jaraco.context==5.3.0
jaraco.functools==4.0.0
jaraco.text==3.12.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.0
jsmin==3.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lesscpy==0.15.1
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==11.0.0
plotly==5.24.1
ply==3.11
portend==3.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2024.1
rcssmin==1.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
scikit-learn==1.4.2
scipy==1.13.0
setuptools==69.2.0
shapely==2.0.6
six==1.16.0
SQLAlchemy==2.0.21
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.13.1
tempora==5.5.1
tenacity==9.0.0
threadpoolctl==3.4.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
traitlets==5.14.3
triton==3.2.0
tsfel==0.1.7
typeguard==4.2.1
typing_extensions==4.13.2
tzdata==2024.1
urllib3==2.2.3
URLObject==2.4.3
visitor==0.1.3
wcwidth==0.2.13
Werkzeug==2.3.7
wheel==0.43.0
WTForms==3.0.1
xxhash==3.5.0
xyzservices==2024.9.0
zc.lockfile==3.0.post1
